BC-9190 - rename `teams` module folder to `team`
