+++ apps/server/src/apps/server.app.ts
@@ -14,7 +14,7 @@ import { InternalServerModule } from '@modules/internal-server/internal-server.a
 import { RocketChatService } from '@modules/rocketchat';
 import { FeathersRosterService } from '@modules/roster';
 import { ServerModule } from '@modules/server/server.app.module';
-import { TeamService } from '@modules/teams';
+import { TeamService } from '@modules/team';
 import { ContextExternalToolService } from '@modules/tool/context-external-tool';
 import { NestFactory } from '@nestjs/core';
 import { ExpressAdapter } from '@nestjs/platform-express';
