+++ apps/server/src/modules/authorization-rules/rules/team.rule.ts
@@ -1,5 +1,5 @@
 import { AuthorizationContext, AuthorizationHelper, AuthorizationInjectionService, Rule } from '@modules/authorization';
-import { TeamEntity, TeamUserEntity } from '@modules/teams/repo';
+import { TeamEntity, TeamUserEntity } from '@modules/team/repo';
 import { User } from '@modules/user/repo';
 import { Injectable } from '@nestjs/common';
