+++ apps/server/src/modules/collaborative-storage/services/collaborative-storage.service.spec.ts
@@ -7,8 +7,8 @@ import { TeamMapper } from '@modules/collaborative-storage/mapper/team.mapper';
 import { CollaborativeStorageService } from '@modules/collaborative-storage/services/collaborative-storage.service';
 import { RoleDto } from '@modules/role/service/dto/role.dto';
 import { RoleService } from '@modules/role/service/role.service';
-import { TeamEntity, TeamsRepo } from '@modules/teams/repo';
-import { teamFactory } from '@modules/teams/testing';
+import { TeamEntity, TeamsRepo } from '@modules/team/repo';
+import { teamFactory } from '@modules/team/testing';
 import { ForbiddenException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { RoleName } from '@shared/domain/interface';
