+++ apps/server/src/modules/deletion/deletion.config.ts
@@ -6,7 +6,7 @@ import { FilesConfig } from '@modules/files';
 import { NewsConfig } from '@modules/news';
 import { PseudonymConfig } from '@modules/pseudonym';
 import { RocketChatUserConfig } from '@modules/rocketchat-user';
-import { TeamsConfig } from '@modules/teams';
+import { TeamsConfig } from '@modules/team';
 
 export interface DeletionConfig
 	extends LoggerConfig,
