+++ apps/server/src/modules/news/mapper/news.mapper.spec.ts
@@ -3,7 +3,7 @@ import { CourseEntity, CourseGroupEntity } from '@modules/course/repo';
 import { courseEntityFactory } from '@modules/course/testing';
 import { SchoolEntity } from '@modules/school/repo';
 import { schoolEntityFactory } from '@modules/school/testing';
-import { TeamEntity } from '@modules/teams/repo';
+import { TeamEntity } from '@modules/team/repo';
 import { User } from '@modules/user/repo';
 import { userFactory } from '@modules/user/testing';
 import { CourseNews, News, NewsProperties, SchoolNews, TeamNews } from '@shared/domain/entity';
