+++ apps/server/src/modules/server/server.app.module.ts
@@ -35,7 +35,7 @@ import { SharingApiModule } from '@modules/sharing/sharing-api.module';
 import { ShdApiModule } from '@modules/shd/shd.api.module';
 import { SystemApiModule } from '@modules/system/system-api.module';
 import { TaskApiModule } from '@modules/task/task-api.module';
-import { TeamsApiModule } from '@modules/teams/teams-api.module';
+import { TeamsApiModule } from '@modules/team/teams-api.module';
 import { ToolApiModule } from '@modules/tool/tool-api.module';
 import { ImportUserModule } from '@modules/user-import';
 import { UserLicenseModule } from '@modules/user-license';
