+++ apps/server/src/modules/video-conference/service/video-conference.service.spec.ts
@@ -15,8 +15,8 @@ import { RoomService } from '@modules/room';
 import { RoomMembershipService } from '@modules/room-membership';
 import { roomMembershipFactory } from '@modules/room-membership/testing';
 import { roomFactory } from '@modules/room/testing';
-import { TeamsRepo } from '@modules/teams/repo';
-import { teamFactory, teamUserFactory } from '@modules/teams/testing';
+import { TeamsRepo } from '@modules/team/repo';
+import { teamFactory, teamUserFactory } from '@modules/team/testing';
 import { UserService } from '@modules/user';
 import { User } from '@modules/user/repo';
 import { userDoFactory, userFactory } from '@modules/user/testing';
