+++ apps/server/src/modules/video-conference/uc/video-conference-deprecated.uc.spec.ts
@@ -6,8 +6,8 @@ import { AuthorizationService } from '@modules/authorization';
 import { CourseService } from '@modules/course';
 import { CourseEntity } from '@modules/course/repo';
 import { LegacySchoolService } from '@modules/legacy-school';
-import { TeamEntity, TeamsRepo } from '@modules/teams/repo';
-import { teamFactory } from '@modules/teams/testing';
+import { TeamEntity, TeamsRepo } from '@modules/team/repo';
+import { teamFactory } from '@modules/team/testing';
 import { UserService } from '@modules/user';
 import { UserDo } from '@modules/user/domain';
 import { User } from '@modules/user/repo';
