+++ apps/server/src/shared/repo/news/news.repo.integration.spec.ts
@@ -1,7 +1,7 @@
 import { NotFoundError } from '@mikro-orm/core';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { CourseEntity, CourseGroupEntity } from '@modules/course/repo';
-import { TeamEntity } from '@modules/teams/repo';
+import { TeamEntity } from '@modules/team/repo';
 import { userFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { CourseNews, News, SchoolNews, TeamNews } from '@shared/domain/entity';
